moved IExtendedDraggable interface to extension package
